move all jobs to their own package
