+++ src/main/java/net/querz/mcaselector/headless/ParamExecutor.java
@@ -8,6 +8,14 @@ import net.querz.mcaselector.exception.ParseException;
 import net.querz.mcaselector.filter.FilterParser;
 import net.querz.mcaselector.filter.GroupFilter;
 import net.querz.mcaselector.io.*;
+import net.querz.mcaselector.io.job.ChunkFilterDeleter;
+import net.querz.mcaselector.io.job.ChunkFilterExporter;
+import net.querz.mcaselector.io.job.ChunkFilterSelector;
+import net.querz.mcaselector.io.job.ChunkImporter;
+import net.querz.mcaselector.io.job.FieldChanger;
+import net.querz.mcaselector.io.job.SelectionDeleter;
+import net.querz.mcaselector.io.job.SelectionExporter;
+import net.querz.mcaselector.io.job.SelectionImageExporter;
 import net.querz.mcaselector.point.Point2i;
 import net.querz.mcaselector.property.DataProperty;
 import net.querz.mcaselector.range.Range;
