+++ src/main/java/net/querz/mcaselector/io/mca/EntitiesMCAFile.java
@@ -1,6 +1,5 @@
 package net.querz.mcaselector.io.mca;
 
-import net.querz.mcaselector.io.CompressionType;
 import net.querz.mcaselector.point.Point2i;
 import net.querz.mcaselector.range.Range;
 import net.querz.nbt.tag.CompoundTag;
