+++ src/main/java/net/querz/mcaselector/ui/DialogHelper.java
@@ -10,6 +10,14 @@ import javafx.stage.FileChooser;
 import javafx.stage.Stage;
 import net.querz.mcaselector.Config;
 import net.querz.mcaselector.io.*;
+import net.querz.mcaselector.io.job.ChunkFilterDeleter;
+import net.querz.mcaselector.io.job.ChunkFilterExporter;
+import net.querz.mcaselector.io.job.ChunkFilterSelector;
+import net.querz.mcaselector.io.job.ChunkImporter;
+import net.querz.mcaselector.io.job.FieldChanger;
+import net.querz.mcaselector.io.job.SelectionDeleter;
+import net.querz.mcaselector.io.job.SelectionExporter;
+import net.querz.mcaselector.io.job.SelectionImageExporter;
 import net.querz.mcaselector.io.mca.ChunkData;
 import net.querz.mcaselector.io.mca.Region;
 import net.querz.mcaselector.tiles.Selection;
